requirements.txt
discord.py
pandas
numpy
